pyproject.toml
[project]
name = "restaurant-booking-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.5.1",
    "duckduckgo-search>=8.1.1",
    "fastapi>=0.131.0",
    "langchain>=1.2.10",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-groq>=1.1.2",
    "langgraph>=1.0.9",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=5.2.3",
    "uvicorn>=0.41.0",
    "huggingface-hub==0.36.2",
    "langchain-huggingface==1.2.0"
]

[dependency-groups]
dev = [
    "black>=26.1.0",
    "pytest>=9.0.2",
    "ruff>=0.15.2",
]
